Cambios para calcular el precio de la entrada según el formato de la sala y la cantidad de entradas

+++ DAL/DAL_ACCESOBD.cs
@@ -78,7 +78,7 @@ namespace DAL
             return param;
         }
 
-        public SqlParameter CrearParametro(string nombre, float valor)
+        public SqlParameter CrearParametro(string nombre, double valor)
         {
             SqlParameter param = new SqlParameter(nombre, valor);
             param.DbType = DbType.Double;
